Orchestrating via LLM & via Code/pyproject.toml
[project]
name = "8thclass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.5",
]
